@@ -28,8 +28,8 @@
     "classnames": "~2.3.1",
     "core-js": "~3.13.1",
     "moment": "~2.29.1",
-    "react": "~16.13.1",
-    "react-dom": "~16.13.1",
+    "react": "~17.0.2",
+    "react-dom": "~17.0.2",
     "react-hot-loader": "~4.13.0",
     "react-modal": "~3.14.2",
     "react-redux": "~7.2.4",
@@ -52,15 +52,15 @@
     "@types/classnames": "~2.3.1",
     "@types/enzyme": "~3.10.8",
     "@types/jest": "~26.0.23",
-    "@types/react": "~16.9.56",
-    "@types/react-dom": "~16.9.12",
+    "@types/react": "~17.0.9",
+    "@types/react-dom": "~17.0.6",
     "@types/react-modal": "~3.12.0",
     "@types/react-redux": "~7.1.16",
     "@types/react-router-dom": "~5.1.7",
-    "@types/react-test-renderer": "~16.9.5",
+    "@types/react-test-renderer": "~17.0.1",
     "@types/redux-mock-store": "~1.0.2",
     "altinn-designsystem": "2.7.12",
-    "babel-jest": "~26.6.3",
+    "babel-jest": "~27.0.2",
     "cross-env": "~7.0.3",
     "css-loader": "~5.2.6",
     "enzyme": "~3.11.0",
@@ -71,7 +71,7 @@
     "fork-ts-checker-webpack-plugin": "~6.2.10",
     "html-loader": "~1.3.2",
     "html-webpack-plugin": "~4.5.2",
-    "jest": "~26.6.3",
+    "jest": "~27.0.4",
     "jest-environment-enzyme": "~7.1.2",
     "jest-enzyme": "~7.1.2",
     "jest-junit": "~12.1.0",
@@ -79,7 +79,7 @@
     "monaco-editor-webpack-plugin": "~3.1.0",
     "prettier": "~2.3.0",
     "prettier-tslint": "~0.4.2",
-    "react-test-renderer": "~16.13.1",
+    "react-test-renderer": "~17.0.2",
     "redux-mock-store": "~1.5.4",
     "resolve-url-loader": "~3.1.3",
     "rimraf": "~3.0.2",
@@ -87,7 +87,7 @@
     "style-loader": "~2.0.0",
     "svg-inline-loader": "~0.8.2",
     "terser-webpack-plugin": "~4.2.3",
-    "ts-jest": "~26.5.6",
+    "ts-jest": "~27.0.3",
     "ts-loader": "~8.1.0",
     "webpack": "~4.44.2",
     "webpack-cli": "~3.3.12",
